+++ applications/fishing-map/src/store.ts
@@ -8,6 +8,8 @@ import {
 import connectedRoutes, { routerQueryMiddleware } from './routes/routes'
 import userReducer from './features/user/user.slice'
 import workspaceReducer from './features/workspace/workspace.slice'
+import datasetsReducer from './features/datasets/datasets.slice'
+import dataviewsReducer from './features/dataviews/dataviews.slice'
 import resourcesReducer from './features/resources/resources.slice'
 import searchReducer from './features/search/search.slice'
 import timebarReducer from './features/timebar/timebar.slice'
@@ -24,6 +26,8 @@ const rootReducer = combineReducers({
   user: userReducer,
   search: searchReducer,
   workspace: workspaceReducer,
+  datasets: datasetsReducer,
+  dataviews: dataviewsReducer,
   resources: resourcesReducer,
   timebar: timebarReducer,
   map: mapReducer,