+++ applications/fishing-map/src/utils/dates.ts
@@ -2,6 +2,7 @@ import { DateTime } from 'luxon'
 
 const DATE_FORMAT = 'yyyy/MM/dd'
 
-export function formatDate(date: string) {
+export function formatDate(date?: string) {
+  if (!date) return ''
   return DateTime.fromISO(date).toFormat(DATE_FORMAT)
 }