+++ packages/dataviews-client/src/index.ts
@@ -1,21 +1,4 @@
 export { default } from './dataviews-client'
 export { default as resolveDataviews } from './resolve-dataviews'
 export { default as resolveEndpoint } from './resolve-endpoint'
-export type {
-  Endpoint,
-  EndpointParam,
-  Dataset,
-  DatasetTypes,
-  DatasetStatus,
-  DatasetConfiguration,
-  Dataview,
-  DataviewCreation,
-  DataviewConfig,
-  DataviewDatasetConfigParams,
-  DataviewDatasetConfig,
-  AOI,
-  Workspace,
-  WorkspaceUpsert,
-  WorkspaceDataviewConfig,
-  Resource,
-} from './types'
+export * from './types'