+++ packages/dataviews-client/src/types.ts
@@ -64,7 +64,7 @@ export interface Dataset {
   createdAt: string
   endpoints?: Endpoint[]
   configuration: DatasetConfiguration | null
-  relatedDatasets: unknown
+  relatedDatasets: Partial<Dataset>[] | null
 }
 
 export interface DataviewConfig {
@@ -99,12 +99,12 @@ export interface DataviewCreation {
 
 export interface Dataview {
   id: number
-  uid?: string // Unique id to identify with datasetConfig params
   name: string
   description: string
   createdAt?: string
   updatedAt?: string
   config: DataviewConfig
+  info?: any
   datasets?: Dataset[]
   datasetsConfig?: DataviewDatasetConfig[]
 }
@@ -117,7 +117,7 @@ export interface AOI {
   bbox: number[]
 }
 
-export interface WorkspaceDataviewConfig {
+export interface DataviewInstance extends Partial<Omit<Dataview, 'id' | 'config'>> {
   id: string
   dataviewId: number
   config?: DataviewConfig
@@ -128,8 +128,6 @@ export interface Workspace {
   id: number
   description: string
   name: string
-  dataviews: Dataview[]
-  dataviewsConfig: WorkspaceDataviewConfig[]
   aoi?: AOI
   viewport: {
     zoom: number
@@ -138,6 +136,9 @@ export interface Workspace {
   }
   start: string
   end: string
+  datasets?: Partial<Dataset>[]
+  dataviews?: Partial<Dataview>[]
+  dataviewInstances: DataviewInstance[]
 }
 
 export interface WorkspaceUpsert extends Partial<Omit<Workspace, 'aoi' | 'dataviews'>> {