+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @globalfishingwatchapp/temporalgrid-demo
 
+## 0.6.1
+
+### Patch Changes
+
+- Updated dependencies [fe37510]
+  - @globalfishingwatch/react-hooks@3.3.0
+
+## 0.6.0
+
+### Minor Changes
+
+- 87dfcf5: animated heatmap legend
+
+### Patch Changes
+
+- Updated dependencies [87dfcf5]
+  - @globalfishingwatch/layer-composer@2.14.0
+  - @globalfishingwatch/react-hooks@3.2.0
+
 ## 0.5.1
 
 ### Patch Changes