+++ packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @globalfishingwatch/react-hooks
 
+## 3.3.0
+
+### Minor Changes
+
+- fe37510: Rename dataviewConfig to dataviewInstance
+
+### Patch Changes
+
+- Updated dependencies [fe37510]
+  - @globalfishingwatch/dataviews-client@3.2.0
+
+## 3.2.0
+
+### Minor Changes
+
+- 87dfcf5: animated heatmap legend
+
+### Patch Changes
+
+- Updated dependencies [87dfcf5]
+  - @globalfishingwatch/layer-composer@2.14.0
+
 ## 3.1.0
 
 ### Minor Changes